ml_pipeline/requirements.txt
influxdb-client==1.38.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0.1
schedule==1.2.0
